refactor: extract CategoryBreakdownSection component
