+++ apps/web-next/src/pages/dashboard/components/CategoryBreakdownSection.tsx
@@ -1,0 +1,94 @@
+import { Card, Col, Row, Table, Typography, type TableColumnsType } from "antd";
+import {
+  TRANSACTION_TYPES,
+  TRANSACTION_TYPE_LABELS,
+  type TransactionType,
+} from "../../../constants/transactionTypes";
+import { type CategorySummary } from "../../../hooks";
+import { useCurrency } from "../../../contexts/currency";
+import { formatCurrency } from "../../../utility/currency";
+
+const { Text, Title } = Typography;
+
+const CategoryBreakdownTable = ({
+  data,
+  loading,
+  type,
+}: {
+  data: CategorySummary[];
+  loading: boolean;
+  type: TransactionType;
+}) => {
+  const { currency } = useCurrency();
+  const filteredData = data.filter((d) => d.type === type);
+
+  const columns: TableColumnsType<CategorySummary> = [
+    {
+      title: "Category",
+      dataIndex: "category_name",
+      key: "category_name",
+      sorter: (a: CategorySummary, b: CategorySummary) =>
+        a.category_name.localeCompare(b.category_name),
+      sortDirections: ["ascend", "descend"],
+    },
+    {
+      title: "Amount",
+      dataIndex: "total",
+      key: "total",
+      sorter: (a: CategorySummary, b: CategorySummary) => a.total - b.total,
+      defaultSortOrder: "descend",
+      sortDirections: ["descend", "ascend"],
+      render: (value: number) => formatCurrency(value, currency),
+      align: "right" as const,
+    },
+  ];
+
+  return (
+    <Table
+      dataSource={filteredData}
+      columns={columns}
+      rowKey={(row) => `${row.type}-${row.category_name}`}
+      loading={loading}
+      pagination={{
+        pageSize: 10,
+        hideOnSinglePage: true,
+        showSizeChanger: false,
+      }}
+      size="small"
+      summary={() => {
+        const total = filteredData.reduce((sum, row) => sum + row.total, 0);
+        return (
+          <Table.Summary.Row>
+            <Table.Summary.Cell index={0}>
+              <Text strong>Total</Text>
+            </Table.Summary.Cell>
+            <Table.Summary.Cell index={1} align="right">
+              <Text strong>{formatCurrency(total, currency)}</Text>
+            </Table.Summary.Cell>
+          </Table.Summary.Row>
+        );
+      }}
+    />
+  );
+};
+
+export const CategoryBreakdownSection = ({
+  data,
+  loading,
+}: {
+  data: CategorySummary[];
+  loading: boolean;
+}) => (
+  <>
+    <Title level={4}>By Category</Title>
+    <Row gutter={[16, 16]}>
+      {Object.values(TRANSACTION_TYPES).map((type) => (
+        <Col xs={24} md={8} key={type}>
+          <Card title={TRANSACTION_TYPE_LABELS[type]} size="small">
+            <CategoryBreakdownTable data={data} loading={loading} type={type} />
+          </Card>
+        </Col>
+      ))}
+    </Row>
+  </>
+);
