+++ apps/web-next/src/pages/dashboard/index.tsx
@@ -1,32 +1,23 @@
 import {
-  Card,
   Col,
   Row,
-  type TableColumnsType,
   Typography,
   Select,
-  Table,
   Tabs,
 } from "antd";
 import { Show } from "@refinedev/antd";
 import { BudgetsSection } from "./BudgetsSection";
 import { ChartsTab } from "./ChartsTab";
 import { TypeSummaryCards } from "./components/TypeSummaryCards";
+import { CategoryBreakdownSection } from "./components/CategoryBreakdownSection";
 
 import { useState, type FC } from "react";
 import dayjs from "dayjs";
-import { useCurrency } from "../../contexts/currency";
-import {
-  TRANSACTION_TYPES,
-  TRANSACTION_TYPE_LABELS,
-  TransactionType,
-} from "../../constants/transactionTypes";
 import {
   usePeriodStats,
-  type CategorySummary,
 } from "../../hooks";
 
-const { Text, Title } = Typography;
+const { Text } = Typography;
 
 // === Constants ===
 const currentYear = dayjs().year();
@@ -40,97 +31,6 @@ const monthOptions = Array.from({ length: 12 }, (_, i) => ({
   label: dayjs().month(i).format("MMMM"),
   value: i,
 }));
-
-// === Utilities ===
-const formatCurrencyLocal = (amount: number, currency: string) =>
-  new Intl.NumberFormat(undefined, {
-    style: "currency",
-    currency,
-  }).format(amount);
-
-// === Components ===
-const CategoryBreakdownTable = ({
-  data,
-  loading,
-  type,
-}: {
-  data: CategorySummary[];
-  loading: boolean;
-  type: TransactionType;
-}) => {
-  const { currency } = useCurrency();
-  const filteredData = data.filter((d) => d.type === type);
-
-  const columns: TableColumnsType<CategorySummary> = [
-    {
-      title: "Category",
-      dataIndex: "category_name",
-      key: "category_name",
-      sorter: (a: CategorySummary, b: CategorySummary) =>
-        a.category_name.localeCompare(b.category_name),
-      sortDirections: ["ascend", "descend"],
-    },
-    {
-      title: "Amount",
-      dataIndex: "total",
-      key: "total",
-      sorter: (a: CategorySummary, b: CategorySummary) => a.total - b.total,
-      defaultSortOrder: "descend",
-      sortDirections: ["descend", "ascend"],
-      render: (value: number) => formatCurrencyLocal(value, currency),
-      align: "right" as const,
-    },
-  ];
-
-  return (
-    <Table
-      dataSource={filteredData}
-      columns={columns}
-      rowKey={(row) => `${row.type}-${row.category_name}`}
-      loading={loading}
-      pagination={{
-        pageSize: 10,
-        hideOnSinglePage: true,
-        showSizeChanger: false,
-      }}
-      size="small"
-      summary={() => {
-        const total = filteredData.reduce((sum, row) => sum + row.total, 0);
-        return (
-          <Table.Summary.Row>
-            <Table.Summary.Cell index={0}>
-              <Text strong>Total</Text>
-            </Table.Summary.Cell>
-            <Table.Summary.Cell index={1} align="right">
-              <Text strong>{formatCurrencyLocal(total, currency)}</Text>
-            </Table.Summary.Cell>
-          </Table.Summary.Row>
-        );
-      }}
-    />
-  );
-};
-
-const CategoryBreakdownSection = ({
-  data,
-  loading,
-}: {
-  data: CategorySummary[];
-  loading: boolean;
-}) => (
-  <>
-    <Title level={4}>By Category</Title>
-    <Row gutter={[16, 16]}>
-      {Object.values(TRANSACTION_TYPES).map((type) => (
-        <Col xs={24} md={8} key={type}>
-          <Card title={TRANSACTION_TYPE_LABELS[type]} size="small">
-            <CategoryBreakdownTable data={data} loading={loading} type={type} />
-          </Card>
-        </Col>
-      ))}
-    </Row>
-  </>
-);
 
 // === Main Component ===
 export const DashboardPage: FC = () => {
